Renamed CRUKDateRange to CRUKSearchkitDateRange.
