+++ index.js
@@ -10,6 +10,6 @@ export { default as CRUKSearchkitLoading } from './src/components/loading/CRUKSe
 export { default as CRUKHitsList } from './src/components/hitsList/HitsList';
 export { default as CRUKSearchHits } from './src/components/hits/Hits';
 export { default as CRUKSearchkitGTM } from './src/components/gtm/CRUKSearchkitGTM';
-export { default as CRUKDateRange } from './src/components/date/CRUKDateRange';
+export { default as CRUKSearchkitDateRange } from './src/components/date/CRUKSearchkitDateRange';
 export { default as CRUKGeoSuggest } from './src/components/geosuggest/CRUKGeoSuggest';
 export { default as CRUKSearchkitSelect } from './src/components/ui/list/CRUKSearchkitSelect';
