+++ src/styles/common.scss
@@ -90,7 +90,7 @@ $react-dates-color-text-focus: #ffffff;
 @import './../components/result/CRUKSearchkitResult';
 @import './../components/summary/CRUKSearchSummary';
 @import './../components/ui/search/CRUKSearchkitSearchUI';
-@import './../components/date/cruk_daterange';
+@import './../components/date/CRUKSearchkitDateRange';
 @import './../components/geosuggest/cruk_geosuggest';
 
 body {
